+++ include/fc/io/raw.hpp
@@ -428,7 +428,7 @@ namespace fc {
        --_max_depth;
        unsigned_int size; fc::raw::unpack( s, size, _max_depth );
        value.clear();
-       value.reserve( std::min( size.value, FC_MAX_PREALLOC_SIZE ) );
+       value.reserve( std::min( size.value, static_cast<uint64_t>(FC_MAX_PREALLOC_SIZE) ) );
        for( uint32_t i = 0; i < size.value; ++i )
        {
           T tmp;
@@ -473,7 +473,7 @@ namespace fc {
        --_max_depth;
        unsigned_int size; fc::raw::unpack( s, size, _max_depth );
        value.clear();
-       value.reserve( std::min( size.value, FC_MAX_PREALLOC_SIZE ) );
+       value.reserve( std::min( size.value, static_cast<uint64_t>(FC_MAX_PREALLOC_SIZE) ) );
        for( uint32_t i = 0; i < size.value; ++i )
        {
           std::pair<K,V> tmp;
@@ -526,11 +526,11 @@ namespace fc {
        FC_ASSERT( _max_depth > 0 );
        --_max_depth;
        unsigned_int size; fc::raw::unpack( s, size, _max_depth );
-       value.resize( std::min( size.value, FC_MAX_PREALLOC_SIZE ) );
+       value.resize( std::min( size.value, static_cast<uint64_t>(FC_MAX_PREALLOC_SIZE) ) );
        for( uint64_t i = 0; i < size; i++ )
        {
           if( i >= value.size() )
-             value.resize( std::min( 2*value.size(), size.value ) );
+             value.resize( std::min( static_cast<uint64_t>(2*value.size()), size.value ) );
           unpack( s, value[i], _max_depth );
        }
     }
@@ -553,11 +553,11 @@ namespace fc {
        FC_ASSERT( _max_depth > 0 );
        --_max_depth;
        unsigned_int size; fc::raw::unpack( s, size, _max_depth );
-       value.resize( std::min( size.value, FC_MAX_PREALLOC_SIZE ) );
+       value.resize( std::min( size.value, static_cast<uint64_t>(FC_MAX_PREALLOC_SIZE) ) );
        for( uint64_t i = 0; i < size; i++ )
        {
           if( i >= value.size() )
-             value.resize( std::min( 2*value.size(), size.value ) );
+             value.resize( std::min( static_cast<uint64_t>(2*value.size()), size.value ) );
           unpack( s, value[i], _max_depth );
        }
     }
